+++ index.d.ts
@@ -1,4 +1,47 @@
-declare const copy: <Value>(value: Value) => Value;
+interface Cache {
+  _keys?: any[];
+  _values?: any[];
+  has: (value: any) => boolean;
+  set: (key: any, value: any) => void;
+  get: (key: any) => any;
+}
 
-export const copyStrict: <Value>(value: Value) => Value;
-export default copy;
+export interface CreateCopierOptions {
+  array?: InternalCopier<any[]>;
+  arrayBuffer?: InternalCopier<ArrayBuffer>;
+  blob?: InternalCopier<Blob>;
+  dataView?: InternalCopier<DataView>;
+  date?: InternalCopier<Date>;
+  map?: InternalCopier<Map<any, any>>;
+  object?: InternalCopier<Record<string, any>>;
+  regExp?: InternalCopier<RegExp>;
+  set?: InternalCopier<Set<any>>;
+}
+
+type InternalCopier<Value = any> = (value: Value, state: State) => Value;
+
+export interface State {
+  Constructor: any;
+  cache: Cache;
+  copier: InternalCopier;
+  prototype: any;
+}
+
+/**
+ * Copy an value deeply as much as possible.
+ */
+export default function copy<Value>(value: Value): Value;
+
+/**
+ * Copy an value deeply as much as possible, where strict recreation of object properties
+ * are maintained. All properties (including non-enumerable ones) are copied with their
+ * original property descriptors on both objects and arrays.
+ */
+export function copyStrict<Value>(value: Value): Value;
+
+/**
+ * Create a custom copier based on the object-specific copy methods passed.
+ */
+export function createCopier(
+  options: CreateCopierOptions
+): <Value>(value: Value) => Value;