+++ src/copier.ts
@@ -2,37 +2,22 @@ import { getCleanClone, getRegExpFlags } from './utils';
 
 import type { Cache } from './utils';
 
-type InternalCopier = <Value = any>(value: Value, cache: Cache) => Value;
+export type InternalCopier<Value = any> = (value: Value, state: State) => Value;
 
-const { isArray } = Array;
+export interface State {
+  Constructor: any;
+  cache: Cache;
+  copier: InternalCopier;
+  prototype: any;
+}
+
 const {
   defineProperty,
   getOwnPropertyDescriptor,
   getOwnPropertyNames,
   getOwnPropertySymbols,
-  getPrototypeOf,
 } = Object;
 const { hasOwnProperty, propertyIsEnumerable } = Object.prototype;
-
-const ARRAY_BUFFER_OBJECT_CLASSES: Record<string, boolean> = {
-  ['[object ArrayBuffer]']: true,
-  ['[object Float32Array]']: true,
-  ['[object Float64Array]']: true,
-  ['[object Int8Array]']: true,
-  ['[object Int16Array]']: true,
-  ['[object Int32Array]']: true,
-  ['[object Uint8Array]']: true,
-  ['[object Uint8ClampedArray]']: true,
-  ['[object Uint16Array]']: true,
-  ['[object Uint32Array]']: true,
-  ['[object Uint64Array]']: true,
-};
-const UNCOPIABLE_OBJECT_CLASSES: Record<string, boolean> = {
-  ['[object Error]']: true,
-  ['[object Promise]']: true,
-  ['[object WeakMap]']: true,
-  ['[object WeakSet]']: true,
-};
 
 const SUPPORTS_SYMBOL = typeof getOwnPropertySymbols === 'function';
 
@@ -46,37 +31,73 @@ const getStrictProperties = SUPPORTS_SYMBOL
   ? getStrictPropertiesModern
   : getOwnPropertyNames;
 
-export function copyArrayLoose(
-  array: any[],
-  prototype: any,
-  copier: InternalCopier,
-  cache: Cache
-) {
-  const clone = new prototype.constructor();
+export function copyArrayLoose(array: any[], state: State) {
+  const clone = new state.Constructor();
+  const { cache, copier } = state;
 
   cache.set(array, clone);
 
   for (let index: number = 0, length = array.length; index < length; ++index) {
-    clone[index] = copier(array[index], cache);
+    clone[index] = copier(array[index], state);
   }
 
   return clone;
 }
 
+export function copyArrayBuffer<Value extends ArrayBuffer>(
+  arrayBuffer: Value,
+  _state: State
+): Value {
+  return arrayBuffer.slice(0) as Value;
+}
+
+export function copyBlob<Value extends Blob>(
+  blob: Value,
+  _state: State
+): Value {
+  return blob.slice(0, blob.size, blob.type) as Value;
+}
+
+export function copyDataView<Value extends DataView>(
+  dataView: Value,
+  state: State
+): Value {
+  return new state.Constructor(dataView.buffer.slice(0));
+}
+
+export function copyDate<Value extends Date>(date: Value, state: State): Value {
+  return new state.Constructor(date.getTime());
+}
+
+export function copyMap<Value extends Map<any, any>>(
+  map: Value,
+  state: State
+): Value {
+  const clone = new state.Constructor() as Value;
+  const { cache, copier } = state;
+
+  cache.set(map, clone);
+
+  map.forEach((value, key) => {
+    clone.set(key, copier(value, state));
+  });
+
+  return map;
+}
+
 function copyObjectLooseLegacy<Value extends {}>(
   object: Value,
-  prototype: any,
-  copier: InternalCopier,
-  cache: Cache
+  state: State
 ): Value {
-  const clone: any = getCleanClone(prototype);
+  const clone: any = getCleanClone(state.prototype);
+  const { cache, copier } = state;
 
   // set in the cache immediately to be able to reuse the object recursively
   cache.set(object, clone);
 
   for (const key in object) {
     if (hasOwnProperty.call(object, key)) {
-      clone[key] = copier(object[key], cache);
+      clone[key] = copier(object[key], state);
     }
   }
 
@@ -85,18 +106,17 @@ function copyObjectLooseLegacy<Value extends {}>(
 
 function copyObjectLooseModern<Value extends {}>(
   object: Value,
-  prototype: any,
-  copier: InternalCopier,
-  cache: Cache
+  state: State
 ): Value {
-  const clone: any = getCleanClone(prototype);
+  const clone: any = getCleanClone(state.prototype);
+  const { cache, copier } = state;
 
   // set in the cache immediately to be able to reuse the object recursively
   cache.set(object, clone);
 
   for (const key in object) {
     if (hasOwnProperty.call(object, key)) {
-      clone[key] = copier(object[key], cache);
+      clone[key] = copier(object[key], state);
     }
   }
 
@@ -110,7 +130,7 @@ function copyObjectLooseModern<Value extends {}>(
     symbol = symbols[index];
 
     if (propertyIsEnumerable.call(object, symbol)) {
-      clone[symbol] = copier((object as any)[symbol], cache);
+      clone[symbol] = copier((object as any)[symbol], state);
     }
   }
 
@@ -131,11 +151,10 @@ export const copyObjectLoose = SUPPORTS_SYMBOL
  */
 export function copyObjectStrict<Value extends {}>(
   object: Value,
-  prototype: any,
-  copier: InternalCopier,
-  cache: Cache
+  state: State
 ): Value {
-  const clone = getCleanClone(prototype);
+  const clone = getCleanClone(state.prototype);
+  const { cache, copier } = state;
 
   // set in the cache immediately to be able to reuse the object recursively
   cache.set(object, clone);
@@ -158,13 +177,13 @@ export function copyObjectStrict<Value extends {}>(
     if (!descriptor) {
       // In extra edge cases where the property descriptor cannot be retrived, fall back to
       // the loose assignment.
-      clone[property] = copier((object as any)[property], cache);
+      clone[property] = copier((object as any)[property], state);
       continue;
     }
 
     // Only clone the value if actually a value, not a getter / setter.
     if (!descriptor.get && !descriptor.set) {
-      descriptor.value = copier(descriptor.value, cache);
+      descriptor.value = copier(descriptor.value, state);
     }
 
     try {
@@ -178,125 +197,32 @@ export function copyObjectStrict<Value extends {}>(
   return clone;
 }
 
-export function copyMap<Value extends Map<any, any>>(
-  value: Value,
-  Constructor: MapConstructor,
-  copier: InternalCopier,
-  cache: Cache
-): Value {
-  const clone = new Constructor() as Value;
-
-  cache.set(value, clone);
-
-  value.forEach((v, k) => {
-    clone.set(k, copier(v, cache));
-  });
-
-  return value;
-}
-
 export function copyRegExp<Value extends RegExp>(
-  value: Value,
-  Constructor: RegExpConstructor
+  regExp: Value,
+  state: State
 ): Value {
-  const clone = new Constructor(value.source, getRegExpFlags(value)) as Value;
+  const clone = new state.Constructor(
+    regExp.source,
+    getRegExpFlags(regExp)
+  ) as Value;
 
-  clone.lastIndex = value.lastIndex;
+  clone.lastIndex = regExp.lastIndex;
 
   return clone;
 }
 
 export function copySet<Value extends Set<any>>(
-  value: Value,
-  Constructor: SetConstructor,
-  copier: InternalCopier,
-  cache: Cache
+  set: Value,
+  state: State
 ): Value {
-  const clone = new Constructor() as Value;
+  const clone = new state.Constructor() as Value;
+  const { cache, copier } = state;
 
-  cache.set(value, clone);
+  cache.set(set, clone);
 
-  value.forEach((v) => {
-    clone.add(copier(v, cache));
+  set.forEach((value) => {
+    clone.add(copier(value, state));
   });
 
-  return value;
-}
-
-export function createCopier(isStrict: boolean) {
-  const copyArray = isStrict ? copyObjectStrict : copyArrayLoose;
-  const copyObject = isStrict ? copyObjectStrict : copyObjectLoose;
-
-  return function copier(value: any, cache: Cache): any {
-    if (!value || typeof value !== 'object') {
-      return value;
-    }
-
-    if (cache.has(value)) {
-      return cache.get(value);
-    }
-
-    const prototype = value.__proto__ || getPrototypeOf(value);
-    const Constructor = prototype && prototype.constructor;
-
-    // plain objects
-    if (!Constructor || Constructor === Object) {
-      return copyObject(value, prototype, copier, cache);
-    }
-
-    // arrays
-    if (isArray(value)) {
-      return copyArray(value, prototype, copier, cache);
-    }
-
-    const objectClass = toString.call(value);
-
-    // dates
-    if (objectClass === '[object Date]') {
-      return new Constructor(value.getTime());
-    }
-
-    // regexps
-    if (objectClass === '[object RegExp]') {
-      return copyRegExp(value, Constructor);
-    }
-
-    // maps
-    if (objectClass === '[object Map]') {
-      return copyMap(value, Constructor, copier, cache);
-    }
-
-    // sets
-    if (objectClass === '[object Set]') {
-      return copySet(value, Constructor, copier, cache);
-    }
-
-    // blobs
-    if (objectClass === '[object Blob]') {
-      return value.slice(0, value.size, value.type);
-    }
-
-    // dataviews
-    if (objectClass === '[object DataView]') {
-      return new Constructor(value.buffer.slice(0));
-    }
-
-    // array buffers
-    if (ARRAY_BUFFER_OBJECT_CLASSES[objectClass]) {
-      return value.slice(0);
-    }
-
-    // if the value cannot / should not be copied deeply, return the reference
-    if (
-      // promise-like
-      typeof value.then === 'function' ||
-      // object classes which cannot be introspected for copy
-      UNCOPIABLE_OBJECT_CLASSES[objectClass]
-    ) {
-      return value;
-    }
-
-    // assume anything left is a custom constructor
-    return copyObject(value, prototype, copier, cache);
-  };
+  return set;
 }